Samples of my work on the C# Project.
This project was an MVC Web App in Visual Studio using the .NET Framework and Entity Framework. The project connected to a database but was Code First which was a unique challenge. We again utilized Azure as the hub to assign stories and manage pull requests. Discord was used for our daily communication this time around. The project was to assist in the creation of a local Theater Companies website. My part was to work on creating and implementing user comments in the "Blog" section of the website. The bulk of coding was done in C# and the Views were done in HTML with C# mixed in using Razor syntax and tags. My work ranged from creating a Comment Model that integrated into the Code First database to the creation of helper methods that truncated text or displayed the time passed since a comment was posted. Examples of my work in this folder.